@ fix: accept fractional ACE wound amountOf in persistence save
ACE fnc_bandageLocal sets a partially-bandaged open wounds amountOf to
(amountOf - impact), a fractional value (e.g. 0.3). CBA encodes it as
"0.3", but WoundEntryConverter read amountOf via GetInt32(), throwing
FormatException and failing the entire persistence_save for any player
with a partially-bandaged wound.

Widen WoundEntry.AmountOf from int to double (it is a continuous quantity
in ACE, not a count) and read it via GetDouble(). classComplex stays int
(10*classID + category, always integer). Legacy docs storing amountOf as
BSON Int32 still load — the driver coerces Int32 to double.

@

+++ UKSF.Api.ArmaServer/Converters/AceMedicalJsonConverters.cs
@@ -31,7 +31,7 @@ public class WoundEntryConverter : JsonConverter<WoundEntry>
         reader.Read();
         var classComplex = reader.GetInt32();
         reader.Read();
-        var amountOf = reader.GetInt32();
+        var amountOf = reader.GetDouble();
         reader.Read();
         var bleedingRate = reader.GetDouble();
         reader.Read();
@@ -58,7 +58,7 @@ public class WoundEntryConverter : JsonConverter<WoundEntry>
             switch (name?.ToLowerInvariant())
             {
                 case "classcomplex": entry.ClassComplex = reader.GetInt32(); break;
-                case "amountof":     entry.AmountOf = reader.GetInt32(); break;
+                case "amountof":     entry.AmountOf = reader.GetDouble(); break;
                 case "bleedingrate": entry.BleedingRate = reader.GetDouble(); break;
                 case "wounddamage":  entry.WoundDamage = reader.GetDouble(); break;
                 default:             reader.Skip(); break;
